+++ modules/querier/querier.go
@@ -21,6 +21,7 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	oteltrace "go.opentelemetry.io/otel/trace"
+	"go.uber.org/atomic"
 	"go.uber.org/multierr"
 	"golang.org/x/sync/semaphore"
 
@@ -81,11 +82,6 @@ type Querier struct {
 
 	subservices        *services.Manager
 	subservicesWatcher *services.FailureWatcher
-}
-
-type responseFromIngesters struct {
-	addr     string
-	response interface{}
 }
 
 type responseFromGenerators struct {
@@ -242,7 +238,6 @@ func (q *Querier) FindTraceByID(ctx context.Context, req *tempopb.TraceByIDReque
 	maxBytes := q.limits.MaxBytesPerTrace(userID)
 	combiner := trace.NewCombiner(maxBytes, req.AllowPartialTrace)
 
-	var spanCount, spanCountTotal, traceCountTotal int
 	if req.QueryMode == QueryModeIngesters || req.QueryMode == QueryModeAll {
 		var getRSFn replicationSetFn
 		if q.cfg.QueryRelevantIngesters {
@@ -251,34 +246,37 @@ func (q *Querier) FindTraceByID(ctx context.Context, req *tempopb.TraceByIDReque
 				return r.Get(traceKey, ring.Read, nil, nil, nil)
 			}
 		}
+		var spanCountTotal, traceCountTotal atomic.Int64
+		var found atomic.Bool
 
 		// get responses from all ingesters in parallel
 		span.AddEvent("searching ingesters")
-		responses, err := q.forIngesterRings(ctx, userID, getRSFn, func(funcCtx context.Context, client tempopb.QuerierClient) (interface{}, error) {
-			return client.FindTraceByID(funcCtx, req)
-		})
+		forEach := func(funcCtx context.Context, client tempopb.QuerierClient) error {
+			resp, err := client.FindTraceByID(funcCtx, req)
+			if err != nil {
+				return err
+			}
+			t := resp.Trace
+			if t != nil {
+				// we found a trace, consume and count it
+				spanCount, err := combiner.Consume(t)
+				if err != nil {
+					return err
+				}
+				spanCountTotal.Add(int64(spanCount))
+				traceCountTotal.Inc()
+				found.Store(true)
+			}
+			return nil
+		}
+		err := q.forIngesterRings(ctx, userID, getRSFn, forEach)
 		if err != nil {
 			return nil, fmt.Errorf("error querying ingesters in Querier.FindTraceByID: %w", err)
 		}
-
-		found := false
-		for _, r := range responses {
-			t := r.response.(*tempopb.TraceByIDResponse).Trace
-			if t != nil {
-				spanCount, err = combiner.Consume(t)
-				if err != nil {
-					return nil, err
-				}
-
-				spanCountTotal += spanCount
-				traceCountTotal++
-				found = true
-			}
-		}
 		span.AddEvent("done searching ingesters", oteltrace.WithAttributes(
-			attribute.Bool("found", found),
-			attribute.Int("combinedSpans", spanCountTotal),
-			attribute.Int("combinedTraces", traceCountTotal)))
+			attribute.Bool("found", found.Load()),
+			attribute.Int64("combinedSpans", spanCountTotal.Load()),
+			attribute.Int64("combinedTraces", traceCountTotal.Load())))
 	}
 
 	if req.QueryMode == QueryModeBlocks || req.QueryMode == QueryModeAll {
@@ -326,23 +324,23 @@ func (q *Querier) FindTraceByID(ctx context.Context, req *tempopb.TraceByIDReque
 }
 
 type (
-	forEachFn        func(ctx context.Context, client tempopb.QuerierClient) (interface{}, error)
+	forEachFn        func(ctx context.Context, client tempopb.QuerierClient) error
 	replicationSetFn func(r ring.ReadRing) (ring.ReplicationSet, error)
 )
 
 // forIngesterRings runs f, in parallel, for given ingesters
-func (q *Querier) forIngesterRings(ctx context.Context, userID string, getReplicationSet replicationSetFn, f forEachFn) ([]responseFromIngesters, error) {
+func (q *Querier) forIngesterRings(ctx context.Context, userID string, getReplicationSet replicationSetFn, f forEachFn) error {
 	if ctx.Err() != nil {
 		_ = level.Debug(log.Logger).Log("forIngesterRings context error", "ctx.Err()", ctx.Err().Error())
-		return nil, ctx.Err()
+		return ctx.Err()
 	}
 
 	// if we have no configured ingester rings this will fail silently. let's return an actual error instead
 	if len(q.ingesterRings) == 0 {
-		return nil, errors.New("forIngesterRings: no ingester rings configured")
+		return errors.New("forIngesterRings: no ingester rings configured")
 	}
 
-	// if a nil replicationsetfn is passed, that means to just use a standard readring
+	// if a nil replicationSetFn is passed, that means to just use a standard Read ring
 	if getReplicationSet == nil {
 		getReplicationSet = func(r ring.ReadRing) (ring.ReplicationSet, error) {
 			return r.GetReplicationSetForOperation(ring.Read)
@@ -352,12 +350,11 @@ func (q *Querier) forIngesterRings(ctx context.Context, userID string, getReplic
 	var mtx sync.Mutex
 	var wg sync.WaitGroup
 
-	var responses []responseFromIngesters
 	var responseErr error
 
-	for i, ring := range q.ingesterRings {
+	for i, ingesterRing := range q.ingesterRings {
 		if q.cfg.ShuffleShardingIngestersEnabled {
-			ring = ring.ShuffleShardWithLookback(
+			ingesterRing = ingesterRing.ShuffleShardWithLookback(
 				userID,
 				q.limits.IngestionTenantShardSize(userID),
 				q.cfg.ShuffleShardingIngestersLookbackPeriod,
@@ -365,17 +362,16 @@ func (q *Querier) forIngesterRings(ctx context.Context, userID string, getReplic
 			)
 		}
 
-		replicationSet, err := getReplicationSet(ring)
+		replicationSet, err := getReplicationSet(ingesterRing)
 		if err != nil {
-			return nil, fmt.Errorf("forIngesterRings: error getting replication set for ring (%d): %w", i, err)
+			return fmt.Errorf("forIngesterRings: error getting replication set for ring (%d): %w", i, err)
 		}
 		pool := q.ingesterPools[i]
 
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			res, err := forOneIngesterRing(ctx, replicationSet, f, pool, q.cfg.ExtraQueryDelay)
-
+			err := forOneIngesterRing(ctx, replicationSet, f, pool, q.cfg.ExtraQueryDelay)
 			mtx.Lock()
 			defer mtx.Unlock()
 
@@ -383,24 +379,20 @@ func (q *Querier) forIngesterRings(ctx context.Context, userID string, getReplic
 				responseErr = multierr.Combine(responseErr, err)
 				return
 			}
-
-			for _, r := range res {
-				responses = append(responses, r.(responseFromIngesters))
-			}
 		}()
 	}
 
 	wg.Wait()
 
 	if responseErr != nil {
-		return nil, responseErr
+		return responseErr
 	}
 
-	return responses, nil
+	return nil
 }
 
-func forOneIngesterRing(ctx context.Context, replicationSet ring.ReplicationSet, f forEachFn, pool *ring_client.Pool, extraQueryDelay time.Duration) ([]interface{}, error) {
-	ctx, span := tracer.Start(ctx, "Querier.forOneIngester")
+func forOneIngesterRing(ctx context.Context, replicationSet ring.ReplicationSet, f forEachFn, pool *ring_client.Pool, extraQueryDelay time.Duration) error {
+	ctx, span := tracer.Start(ctx, "Querier.forOneIngesterRing")
 	defer span.End()
 
 	doFunc := func(funcCtx context.Context, ingester *ring.InstanceDesc) (interface{}, error) {
@@ -414,15 +406,22 @@ func forOneIngesterRing(ctx context.Context, replicationSet ring.ReplicationSet,
 			return nil, fmt.Errorf("failed to get client for %s: %w", ingester.Addr, err)
 		}
 
-		resp, err := f(funcCtx, client.(tempopb.QuerierClient))
+		err = f(funcCtx, client.(tempopb.QuerierClient))
 		if err != nil {
 			return nil, fmt.Errorf("failed to execute f() for %s: %w", ingester.Addr, err)
 		}
 
-		return responseFromIngesters{ingester.Addr, resp}, nil
+		// we are returning the empty response here because response is collected by
+		// the collector inside forEachFn
+		return nil, nil
 	}
 
-	return replicationSet.Do(ctx, extraQueryDelay, doFunc)
+	// ignore response because it's nil, and we are using a collector inside forEachFn to
+	// collect the actual response. we need to return nil here and ignore it
+	// because doFunc expects us to return a response
+	_, err := replicationSet.Do(ctx, extraQueryDelay, doFunc)
+
+	return err
 }
 
 // forGivenGenerators runs f, in parallel, for given generators
@@ -477,14 +476,25 @@ func (q *Querier) SearchRecent(ctx context.Context, req *tempopb.SearchRequest) 
 		return nil, fmt.Errorf("error extracting org id in Querier.Search: %w", err)
 	}
 
-	responses, err := q.forIngesterRings(ctx, userID, nil, func(ctx context.Context, client tempopb.QuerierClient) (interface{}, error) {
-		return client.SearchRecent(ctx, req)
-	})
+	var results []*tempopb.SearchResponse
+	mtx := sync.Mutex{}
+
+	forEach := func(ctx context.Context, client tempopb.QuerierClient) error {
+		resp, err := client.SearchRecent(ctx, req)
+		if err != nil {
+			return err
+		}
+		mtx.Lock()
+		defer mtx.Unlock()
+		results = append(results, resp)
+		return nil
+	}
+	err = q.forIngesterRings(ctx, userID, nil, forEach)
 	if err != nil {
 		return nil, fmt.Errorf("error querying ingesters in Querier.Search: %w", err)
 	}
 
-	return q.postProcessIngesterSearchResults(req, responses), nil
+	return q.postProcessIngesterSearchResults(req, results), nil
 }
 
 func (q *Querier) SearchTagsBlocks(ctx context.Context, req *tempopb.SearchTagsBlockRequest) (*tempopb.SearchTagsResponse, error) {
@@ -536,21 +546,22 @@ func (q *Querier) SearchTags(ctx context.Context, req *tempopb.SearchTagsRequest
 	limit := q.limits.MaxBytesPerTagValuesQuery(userID)
 	distinctValues := collector.NewDistinctString(limit)
 
-	lookupResults, err := q.forIngesterRings(ctx, userID, nil, func(ctx context.Context, client tempopb.QuerierClient) (interface{}, error) {
-		return client.SearchTags(ctx, req)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTags: %w", err)
-	}
-
-outerLoop:
-	for _, resp := range lookupResults {
-		for _, res := range resp.response.(*tempopb.SearchTagsResponse).TagNames {
-			distinctValues.Collect(res)
+	forEach := func(ctx context.Context, client tempopb.QuerierClient) error {
+		resp, err := client.SearchTags(ctx, req)
+		if err != nil {
+			return err
+		}
+		for _, tag := range resp.TagNames {
+			distinctValues.Collect(tag)
 			if distinctValues.Exceeded() {
-				break outerLoop // break out of all loops
+				break // stop early
 			}
 		}
+		return nil
+	}
+	err = q.forIngesterRings(ctx, userID, nil, forEach)
+	if err != nil {
+		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTags: %w", err)
 	}
 
 	if distinctValues.Exceeded() {
@@ -570,27 +581,29 @@ func (q *Querier) SearchTagsV2(ctx context.Context, req *tempopb.SearchTagsReque
 		return nil, fmt.Errorf("error extracting org id in Querier.SearchTags: %w", err)
 	}
 
-	// Get results from all ingesters
-	lookupResults, err := q.forIngesterRings(ctx, userID, nil, func(ctx context.Context, client tempopb.QuerierClient) (interface{}, error) {
-		return client.SearchTagsV2(ctx, req)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTags: %w", err)
-	}
-
 	limit := q.limits.MaxBytesPerTagValuesQuery(userID)
 	distinctValues := collector.NewScopedDistinctString(limit)
 
-outerLoop:
-	for _, resp := range lookupResults {
-		for _, res := range resp.response.(*tempopb.SearchTagsV2Response).Scopes {
+	// Get results from all ingesters
+	forEach := func(ctx context.Context, client tempopb.QuerierClient) error {
+		resp, err := client.SearchTagsV2(ctx, req)
+		if err != nil {
+			return err
+		}
+		for _, res := range resp.Scopes {
 			for _, tag := range res.Tags {
 				distinctValues.Collect(res.Name, tag)
 				if distinctValues.Exceeded() {
-					break outerLoop // break out of all loops
+					return nil
 				}
 			}
 		}
+		return nil
+	}
+
+	err = q.forIngesterRings(ctx, userID, nil, forEach)
+	if err != nil {
+		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTags: %w", err)
 	}
 
 	if distinctValues.Exceeded() {
@@ -626,21 +639,24 @@ func (q *Querier) SearchTagValues(ctx context.Context, req *tempopb.SearchTagVal
 		distinctValues.Collect(v)
 	}
 
-	lookupResults, err := q.forIngesterRings(ctx, userID, nil, func(ctx context.Context, client tempopb.QuerierClient) (interface{}, error) {
-		return client.SearchTagValues(ctx, req)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTagValues: %w", err)
-	}
-
-outerLoop:
-	for _, resp := range lookupResults {
-		for _, res := range resp.response.(*tempopb.SearchTagValuesResponse).TagValues {
+	forEach := func(ctx context.Context, client tempopb.QuerierClient) error {
+		resp, err := client.SearchTagValues(ctx, req)
+		if err != nil {
+			return err
+		}
+		for _, res := range resp.TagValues {
 			distinctValues.Collect(res)
 			if distinctValues.Exceeded() {
-				break outerLoop // break out of all loops
+				return nil
 			}
 		}
+
+		return nil
+	}
+
+	err = q.forIngesterRings(ctx, userID, nil, forEach)
+	if err != nil {
+		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTagValues: %w", err)
 	}
 
 	if distinctValues.Exceeded() {
@@ -676,21 +692,22 @@ func (q *Querier) SearchTagValuesV2(ctx context.Context, req *tempopb.SearchTagV
 		return valuesToV2Response(distinctValues), nil
 	}
 
-	// Get results from all ingesters
-	lookupResults, err := q.forIngesterRings(ctx, userID, nil, func(ctx context.Context, client tempopb.QuerierClient) (interface{}, error) {
-		return client.SearchTagValuesV2(ctx, req)
-	})
-	if err != nil {
-		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTagValues: %w", err)
-	}
-
-outerLoop:
-	for _, resp := range lookupResults {
-		for _, res := range resp.response.(*tempopb.SearchTagValuesV2Response).TagValues {
-			if distinctValues.Collect(*res) {
-				break outerLoop // break out of all loops
+	forEach := func(ctx context.Context, client tempopb.QuerierClient) error {
+		resp, err := client.SearchTagValuesV2(ctx, req)
+		if err != nil {
+			return err
+		}
+		for _, res := range resp.TagValues {
+			distinctValues.Collect(*res)
+			if distinctValues.Exceeded() {
+				break // stop early
 			}
 		}
+		return nil
+	}
+	err = q.forIngesterRings(ctx, userID, nil, forEach)
+	if err != nil {
+		return nil, fmt.Errorf("error querying ingesters in Querier.SearchTagValues: %w", err)
 	}
 
 	if distinctValues.Exceeded() {
@@ -1084,15 +1101,14 @@ func (q *Querier) internalTagValuesSearchBlockV2(ctx context.Context, req *tempo
 	return valuesToV2Response(valueCollector), nil
 }
 
-func (q *Querier) postProcessIngesterSearchResults(req *tempopb.SearchRequest, rr []responseFromIngesters) *tempopb.SearchResponse {
+func (q *Querier) postProcessIngesterSearchResults(req *tempopb.SearchRequest, results []*tempopb.SearchResponse) *tempopb.SearchResponse {
 	response := &tempopb.SearchResponse{
 		Metrics: &tempopb.SearchMetrics{},
 	}
 
 	traces := map[string]*tempopb.TraceSearchMetadata{}
 
-	for _, r := range rr {
-		sr := r.response.(*tempopb.SearchResponse)
+	for _, sr := range results {
 		for _, t := range sr.Traces {
 			// Just simply take first result for each trace
 			if _, ok := traces[t.TraceID]; !ok {